[남민정] AnimalFile코드 구현 완료
- 파일 생성 후 getAnimalOne으로 갈 수 있게 코드 구현.

  service
- AnimalFileService.java
  -> addAnimalFile에 파일 저장, 경로, 확장자, 파일이름, 저장 위치 코드 구현

  controller
- AnimalFileController.java
  -> addAnimalFile에 animalFile, animalId 불러옴.

+++ OBO/src/main/java/com/gd/obo/vo/AnimalFile.java
@@ -7,7 +7,7 @@ public class AnimalFile {
 	int animalFileId;
 	int animalId;
 	String animalFileName;
-	int animalFileSize;
+	double animalFileSize; // multipartFile 때문에 double로 바꿈
 	String animalFileExt;
 	String animalFileDate;
 }
